# openapi/app/models.py
# generated by fastapi-codegen:
#   filename:  ./openapi/creatures.oas.0.0.yaml
#   timestamp: 2024-10-27T07:09:28+00:00

# Issues
# Generates example and not examples in each field property which is not a valid option
# Doesn't generate multiple examples
# Generates constr and conint which don't seem to work
# or is this just an error in trunk and not a real one - https://github.com/microsoft/pylance-release/issues/5457

from __future__ import annotations

from enum import Enum
from typing import Optional

from pydantic import BaseModel, Field, conint, constr


class Type(Enum):
    Werewolf = 'Werewolf'
    Dragon = 'Dragon'
    Vampire = 'Vampire'
    Demon = 'Demon'
    Other = 'Other'


class Creature(BaseModel):
    id: Optional[str] = Field(
        None, description='The unique identifier for the creature', examples=['1']
    )
    name: str = Field(
        ..., description='The name of the creature',min_length=1, max_length=20, examples=['Fenrir']
    )
    type: Type = Field(
        ..., description='The type or category of the creature', examples=['Wolf']
    )
    age: Optional[conint(ge=0, le=1000)] = Field(
        None, description='The age of the creature', examples=[120]
    )